Move logic to SecretsStore and test it there
