+++ src/Aspire.Hosting/DistributedApplicationBuilder.cs
@@ -265,7 +265,11 @@ public class DistributedApplicationBuilder : IDistributedApplicationBuilder
                 if (!IsDashboardUnsecured(_innerBuilder.Configuration))
                 {
                     // Passed to apps as a standard OTEL attribute to include in OTLP requests and the dashboard to validate.
-                    EnsureOtlpApiKey(_innerBuilder.Configuration, AppHostAssembly);
+                    // Set a random API key for the OTLP exporter if one isn't already present in configuration.
+                    // If a key is generated, it's stored in the user secrets store so that it will be auto-loaded
+                    // on subsequent runs and not recreated. This is important to ensure it doesn't change the state
+                    // of persistent containers (as a new key would be a spec change).
+                    SecretsStore.GetOrSetUserSecret(_innerBuilder.Configuration, AppHostAssembly, "AppHost:OtlpApiKey", TokenGenerator.GenerateToken);
 
                     // Determine the frontend browser token.
                     if (_innerBuilder.Configuration[KnownConfigNames.DashboardFrontendBrowserToken] is not { Length: > 0 } browserToken)
@@ -366,33 +370,6 @@ public class DistributedApplicationBuilder : IDistributedApplicationBuilder
 
         _innerBuilder.Services.AddSingleton(ExecutionContext);
         LogBuilderConstructed(this);
-    }
-
-    private static void EnsureOtlpApiKey(ConfigurationManager configuration, Assembly? appHostAssembly)
-    {
-        // Set a random API key for the OTLP exporter if one isn't already present in configuration.
-        // If a key is generated, it's stored in the user secrets store so that it will be auto-loaded
-        // on subsequent runs and not recreated. This is important to ensure it doesn't change the state
-        // of persistent containers (as a new key would be a spec change).
-        var key = "AppHost:OtlpApiKey";
-        var existingValue = configuration[key];
-        if (existingValue is null)
-        {
-            // Create the key and store in user secrets
-            var apiKey = TokenGenerator.GenerateToken();
-            configuration.AddInMemoryCollection(
-                new Dictionary<string, string?>
-                {
-                    [key] = apiKey
-                }
-            );
-            if (!SecretsStore.TrySetUserSecret(appHostAssembly, key, apiKey))
-            {
-                // This is a best-effort operation, so we don't throw if it fails. Common reason for failure is that the user secrets ID is not set
-                // in the application's assembly. Note there's no ILogger available this early in the application lifecycle.
-                Debug.WriteLine($"Failed to save OTLP API key in application to user secrets.");
-            }
-        }
     }
 
     private void ConfigureHealthChecks()
