+++ tests/Aspire.Hosting.Tests/DistributedApplicationBuilderTests.cs
@@ -1,19 +1,14 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Reflection;
-using System.Reflection.Emit;
-using System.Runtime.Loader;
 using Aspire.Hosting.Dashboard;
 using Aspire.Hosting.Dcp;
 using Aspire.Hosting.Devcontainers;
 using Aspire.Hosting.Lifecycle;
 using Aspire.Hosting.Publishing;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Configuration.UserSecrets;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using Microsoft.Extensions.SecretManager.Tools.Internal;
 using Xunit;
 
 namespace Aspire.Hosting.Tests;
@@ -95,53 +90,6 @@ public class DistributedApplicationBuilderTests
         Assert.Equal("/path/", publishOptions.Value.OutputPath);
     }
 
-    private static readonly ConstructorInfo s_userSecretsIdAttrCtor = typeof(UserSecretsIdAttribute).GetConstructor([typeof(string)])!;
-
-    [Fact]
-    public void BuilderSavesGeneratedOtlpApiKeyToUserSecrets()
-    {
-        var userSecretsId = Guid.NewGuid().ToString("N");
-        ClearUsersSecrets(userSecretsId);
-
-        var testAssembly = AssemblyBuilder.DefineDynamicAssembly(
-            new("testhost"), AssemblyBuilderAccess.RunAndCollect, [new CustomAttributeBuilder(s_userSecretsIdAttrCtor, [userSecretsId])]);
-
-        var args = new[] { "--environment", "Development" };
-        var builder = DistributedApplication.CreateBuilder(new DistributedApplicationOptions { Args = args, DisableDashboard = false, AssemblyName = testAssembly.FullName });
-        var userSecrets = GetUserSecrets(userSecretsId);
-
-        var configValue = builder.Configuration["AppHost:OtlpApiKey"];
-        Assert.True(userSecrets.TryGetValue("AppHost:OtlpApiKey", out var savedValue));
-        Assert.Equal(configValue, savedValue);
-
-        DeleteUserSecretsFile(userSecretsId);
-    }
-
-    [Fact]
-    public void BuilderReadsOtlpApiKeyFromUserSecrets()
-    {
-        var userSecretsId = Guid.NewGuid().ToString("N");
-        ClearUsersSecrets(userSecretsId);
-
-        var testAssembly = AssemblyBuilder.DefineDynamicAssembly(
-            new("forthetest"), AssemblyBuilderAccess.RunAndCollect, [new CustomAttributeBuilder(s_userSecretsIdAttrCtor, [userSecretsId])]);
-
-        Assembly? resolving(AssemblyLoadContext context, AssemblyName name) => name.FullName == testAssembly.FullName ? testAssembly : null;
-
-        AssemblyLoadContext.Default.Resolving += resolving;
-
-        var otlpApiKey = TokenGenerator.GenerateToken();
-        Assert.True(SecretsStore.TrySetUserSecret(testAssembly, "AppHost:OtlpApiKey", otlpApiKey));
-
-        var args = new[] { "--environment", "Development" };
-        var builder = DistributedApplication.CreateBuilder(new DistributedApplicationOptions { Args = args, DisableDashboard = false, AssemblyName = testAssembly.FullName });
-
-        Assert.Equal(otlpApiKey, builder.Configuration["AppHost:OtlpApiKey"]);
-
-        DeleteUserSecretsFile(userSecretsId);
-        AssemblyLoadContext.Default.Resolving -= resolving;
-    }
-
     [Fact]
     public void AppHostDirectoryAvailableViaConfig()
     {
@@ -221,32 +169,11 @@ public class DistributedApplicationBuilderTests
         Assert.Equal("Multiple resources with the name 'Test'. Resource names are case-insensitive.", ex.Message);
     }
 
-    private static Dictionary<string, string?> GetUserSecrets(string userSecretsId)
-    {
-        var secretsStore = new SecretsStore(userSecretsId);
-        return secretsStore.AsEnumerable().ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
-    }
-
-    private static void ClearUsersSecrets(string userSecretsId)
-    {
-        var secretsStore = new SecretsStore(userSecretsId);
-        secretsStore.Clear();
-        secretsStore.Save();
-    }
-
-    private static void DeleteUserSecretsFile(string userSecretsId)
-    {
-        var userSecretsPath = PathHelper.GetSecretsPathFromSecretsId(userSecretsId);
-        if (File.Exists(userSecretsPath))
-        {
-            File.Delete(userSecretsPath);
-        }
-    }
-
     private sealed class TestResource : IResource
     {
         public string Name => nameof(TestResource);
 
         public ResourceAnnotationCollection Annotations => throw new NotImplementedException();
     }
+
 }
